feat: use page.routeBase.name instead of page.name
